+++ recipes/plain.v
@@ -4,11 +4,14 @@
    files ‘--frontend coq+rst’, but you can use ‘--frontend coq’ to skip all
    comment processing:
 
-      alectryon --frontend coq plain.v # Coq → HTML; produces ‘plain.v.html’
+      alectryon --frontend coq plain.v
+          # Coq → HTML; produces ‘plain.v.html’
+      alectryon --frontend coq --body-only plain.v -o plain.v.body.html
+          # Coq → HTML; produces ‘plain.v.body.html’
 
    Conversion from plain Coq to LaTeX isn't supported yet (use coq+rst):
 
-      # alectryon --frontend coq --backend latex plain.v # Coq → LaTeX; produces ‘plain.v.tex’ *)
+      # alectryon --frontend coq --backend latex plain.v # XXX: not supported *)
 
 Require Import List.